Dropped host_restrictions and moved desktop launcher to different
playbook.
This was done to help make this role executable without requiring
prior manual edits when tested by others than me. The host_restrictions
var was an idea in the same vein, but testing has proved it did
not work as intended.
Better to simply drop the creation of desktop launcher which
was the root of the problem. It was always an unstable foundation
to attempt to incorporate both the desktop and the server
component in the same role.